Add files via upload
This program displays each of the six colors of the rainbow with their congruent colors consecutively followed by 10 iterations of a random word out of the six colors of the rainbow with a random incongruent color. Each word and color display is shown for 3 seconds and is immediately followed by the next word and color display.