Create graph.m
In this part of the project, we are trying to graph the eye movements of the user between the time the word and color stimulus displays and the time they press the key. We will have a total of 24 graphs (12 for each word and color display and 2 for each eye). The eye tracker produces an EDF file with we convert to an .asc file. The .asc file records the time of recording in column 1, the left eye x position in column 2, the left eye y position in column 3, the right eye x position in column 5, and the right eye y position in column 6. Using the reaction time data in the .dat file and the fact that the fixation point displays for 1 second before each word and color display, this program determines the time ranges for the recordings in the .asc file that need to be graphed.